NowPlaying/requirements.txt
PyYAML~=5.3.1
wxPython~=4.1.0
spotipy~=2.16.1
pyinstaller~=4.0